requirements.txt
opencv-python
numpy
pygame
scikit-learn
Pillow
is-wire
is-msgs
protobuf
six
